requirements.txt
fastapi
uvicorn
ultralytics
easyocr
pillow
python-multipart
numpy
